pyproject.toml
[tool.poetry]
name = "adalayers"
version = "0.1.0"
description = "adaptive layers selection for transfer learning"
authors = ["alexdremov <dremov.me@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
numpy = "^1.26.2"
wandb = "^0.16.0"
torch = [
    { markers = "sys_platform == 'darwin' and platform_machine == 'arm64'", url = "https://download.pytorch.org/whl/cpu/torch-2.1.1-cp311-none-macosx_11_0_arm64.whl#sha256=e312f7e82e49565f7667b0bbf9559ab0c597063d93044740781c02acd5a87978" },
    { markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.1.1-cp311-none-macosx_10_9_x86_64.whl#sha256=a70593806f1d7e6b53657d96810518da0f88ef2608c98a402955765b8c79d52c" },
    { markers = "sys_platform == 'linux'", url = "https://download.pytorch.org/whl/cu118/torch-2.1.1%2Bcu118-cp311-cp311-linux_x86_64.whl#sha256=f3c0ba02b50d0021ff26f030e22d4c45965537cf91f322e52a65b8c58396f81c" },
]
torchvision = "^0.16.1"
torchdata = "^0.7.1"
transformers = "^4.35.2"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
hydra-colorlog = "^1.2.0"
datasets = "^2.16.1"
lightning = "^2.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
